refactor(core): split search cache modules
