+++ packages/core/src/discovery/__tests__/cache.test.ts
@@ -14,7 +14,6 @@ import {
   loadCachedSessions,
   markAgentCacheInitialized,
   markAgentFullSyncCompleted,
-  parseSearchQuery,
   searchFileActivitySessions,
   searchSessions,
   saveCachedSessionChanges,
@@ -564,28 +563,6 @@ describe("getCacheInfo", () => {
 });
 
 describe("searchSessions", () => {
-  it("parses lightweight structured search qualifiers", () => {
-    expect(
-      parseSearchQuery(
-        'agent:codex project:"code sesh" projectkind:git_remote projectkey:github.com/acme/app tag:feature-dev tool:apply_patch file:"src/App File.tsx" cost:>1 needle',
-      ),
-    ).toEqual({
-      text: "needle",
-      filters: {
-        agent: "codex",
-        project: "code sesh",
-        projectKind: "git_remote",
-        projectKey: "github.com/acme/app",
-        tags: ["feature-dev"],
-        tools: ["apply_patch"],
-        file: "src/App File.tsx",
-        costMin: 1,
-        costMinExclusive: true,
-      },
-      hasQualifiers: true,
-    });
-  });
-
   it("preserves strict cost qualifier comparisons", () => {
     const below = {
       ...makeSession("below"),
